feat(Report Problem): move shouldShowReportProblem to server and consolidate tests in one file
