+++ app/routes/shared/components/FormFlowPage.tsx
@@ -28,6 +28,7 @@ export function FormFlowPage() {
     translations,
     validFlowPaths,
     flowId,
+    shouldShowReportProblem,
   } = useLoaderData<typeof loader>();
 
   const formFlowMemo = useMemo(
@@ -39,10 +40,6 @@ export function FormFlowPage() {
     }),
     [prunedUserData, validFlowPaths, translations, flowId],
   );
-
-  const showReportProblem =
-    flowId === "/beratungshilfe/antrag" ||
-    flowId === "/prozesskostenhilfe/formular";
 
   useFocusFirstH1();
 
@@ -92,7 +89,7 @@ export function FormFlowPage() {
             <PageContent content={postFormContent} fullScreen={false} />
           </div>
         </div>
-        {showReportProblem && <ReportProblem />}
+        {shouldShowReportProblem && <ReportProblem />}
       </div>
     </FormFlowContext.Provider>
   );
